[DRAFT] Pull up groupMappings
